refactor: combine slide and slideAction stores
