+++ frontend/src/components/SlideNavigationPanel.vue
@@ -66,8 +66,7 @@ import { call } from 'frappe-ui'
 import Draggable from 'vuedraggable'
 
 import { presentation } from '@/stores/presentation'
-import { slideIndex } from '@/stores/slide'
-import { changeSlide, insertSlide } from '@/stores/slideActions'
+import { slideIndex, changeSlide, insertSlide } from '@/stores/slide'
 
 const showNavigator = defineModel('showNavigator', {
 	type: Boolean,
